Added new file access functions. Added a function to save the loaded module from memory, for debugging purpose. Added a memory padding macro for Doynaxdepack. Added several graphics function that can work on the bitplan level. Added a routine that displays the amount of free memory. Added a temporary blitter-based scrolling function. Replaced malloc() calls to calloc() calls. Replaced a struct by a typedef in the sound routines. Simpler but more stable color fade in/fade out. Fixed a memory de-allocation issue when removing a sound. Fixed a bug in the ADPCM 16bit -> 8bit conversion. Minimal documentation effort.

+++ rpage/easing.c
@@ -70,7 +70,7 @@ const UWORD easing_table[] =
 	1023, 	1023, 	1023,  
 };
 
-const int sintab32[1024] =
+const short sintab32[1024] =
 {
     0, 6, 13, 19, 25, 31, 38, 44,
     50, 57, 63, 69, 75, 82, 88, 94,
